# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kabadiwala Connect API",
    "description": "Vernacular, low-literacy offline-tolerant platform connecting informal e-waste collectors with authorized recyclers.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/materials": {
      "get": {
        "tags": [
          "Materials"
        ],
        "summary": "List materials catalog",
        "operationId": "list_materials_api_v1_materials_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by broad category",
              "title": "Category"
            },
            "description": "Filter by broad category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialResponse"
                  },
                  "title": "Response List Materials Api V1 Materials Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Add material to catalog (Admin only)",
        "operationId": "create_material_api_v1_materials_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/materials/{material_id}": {
      "get": {
        "tags": [
          "Materials"
        ],
        "summary": "Get material by ID",
        "operationId": "get_material_api_v1_materials__material_id__get",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Material Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Materials"
        ],
        "summary": "Update material entry (Admin only)",
        "operationId": "update_material_api_v1_materials__material_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Material Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/prices": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "List recorded scrap benchmark prices",
        "operationId": "list_prices_api_v1_prices_get",
        "parameters": [
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by district",
              "title": "District"
            },
            "description": "Filter by district"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category",
              "title": "Category"
            },
            "description": "Filter by category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceResponse"
                  },
                  "title": "Response List Prices Api V1 Prices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prices"
        ],
        "summary": "Record a new benchmark price quote",
        "operationId": "create_price_api_v1_prices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prices/board": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Get active price board for district",
        "operationId": "get_price_board_api_v1_prices_board_get",
        "parameters": [
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "District name, e.g. Mumbai, Pune",
              "default": "Mumbai Suburban",
              "title": "District"
            },
            "description": "District name, e.g. Mumbai, Pune"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceBoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prices/{price_id}": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Get price record by ID",
        "operationId": "get_price_api_v1_prices__price_id__get",
        "parameters": [
          {
            "name": "price_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Price Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots": {
      "get": {
        "tags": [
          "Lots & Transactions"
        ],
        "summary": "List e-waste lots",
        "operationId": "list_lots_api_v1_lots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by transaction status",
              "title": "Status"
            },
            "description": "Filter by transaction status"
          },
          {
            "name": "collector_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by collector",
              "title": "Collector Id"
            },
            "description": "Filter by collector"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LotResponse"
                  },
                  "title": "Response List Lots Api V1 Lots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lots & Transactions"
        ],
        "summary": "Create e-waste lot (idempotent via client_lot_id)",
        "operationId": "create_lot_api_v1_lots_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}": {
      "get": {
        "tags": [
          "Lots & Transactions"
        ],
        "summary": "Get lot details by ID",
        "operationId": "get_lot_api_v1_lots__lot_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}/status": {
      "patch": {
        "tags": [
          "Lots & Transactions"
        ],
        "summary": "Update lot state machine status",
        "operationId": "update_lot_status_api_v1_lots__lot_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LotCreate": {
        "properties": {
          "client_lot_id": {
            "type": "string",
            "title": "Client Lot Id",
            "description": "Client-generated unique UUID for offline idempotency"
          },
          "category": {
            "type": "string",
            "maxLength": 50,
            "title": "Category"
          },
          "weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg"
          },
          "quoted_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Quoted Price"
          },
          "collection_lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Collection Lat"
          },
          "collection_lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Collection Lng"
          },
          "created_at_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At Utc"
          },
          "photo_hashes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Hashes"
          }
        },
        "type": "object",
        "required": [
          "client_lot_id",
          "category",
          "weight_kg",
          "quoted_price",
          "collection_lat",
          "collection_lng"
        ],
        "title": "LotCreate"
      },
...
      "MaterialCreate": {
        "properties": {
          "category": {
            "type": "string",
            "maxLength": 50,
            "title": "Category"
          },
          "sub_category": {
            "type": "string",
            "maxLength": 50,
            "title": "Sub Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "image_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Ref"
          },
          "approx_weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Approx Weight Kg"
          },
          "condition": {
            "type": "string",
            "pattern": "^(working|broken|damaged|burnt)$",
            "title": "Condition",
            "default": "broken"
          },
          "source_type": {
            "type": "string",
            "pattern": "^(household|office|shop|repair_unit|other)$",
            "title": "Source Type",
            "default": "household"
          },
          "estimated_value": {
            "type": "number",
            "minimum": 0.0,
            "title": "Estimated Value"
          }
        },
        "type": "object",
        "required": [
          "category",
          "sub_category",
          "description",
          "approx_weight_kg",
          "estimated_value"
        ],
        "title": "MaterialCreate"
      },
      "MaterialResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "sub_category": {
            "type": "string",
            "title": "Sub Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "image_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Ref"
          },
          "approx_weight_kg": {
            "type": "number",
            "title": "Approx Weight Kg"
          },
          "condition": {
            "type": "string",
            "title": "Condition"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "estimated_value": {
            "type": "number",
            "title": "Estimated Value"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "sub_category",
          "description",
          "image_ref",
          "approx_weight_kg",
          "condition",
          "source_type",
          "estimated_value",
          "created_at",
          "updated_at"
        ],
        "title": "MaterialResponse"
      },
      "MaterialUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "approx_weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Approx Weight Kg"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(working|broken|damaged|burnt)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(household|office|shop|repair_unit|other)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "estimated_value": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Value"
          }
        },
        "type": "object",
        "title": "MaterialUpdate"
      },
...
      "PriceBoardResponse": {
        "properties": {
          "district": {
            "type": "string",
            "title": "District"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          },
          "rates": {
            "items": {
              "$ref": "#/components/schemas/PriceBoardCategory"
            },
            "type": "array",
            "title": "Rates"
          }
        },
        "type": "object",
        "required": [
          "district",
          "valid_until",
          "rates"
        ],
        "title": "PriceBoardResponse"
      },
      "PriceCreate": {
        "properties": {
          "category": {
            "type": "string",
            "maxLength": 50,
            "title": "Category"
          },
          "sub_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Category"
          },
          "district": {
            "type": "string",
            "maxLength": 100,
            "title": "District"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "buying_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Buying Price"
          },
          "selling_quoted_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Selling Quoted Price"
          },
          "unit": {
            "type": "string",
            "pattern": "^(kg|piece)$",
            "title": "Unit",
            "default": "kg"
          },
          "market_min": {
            "type": "number",
            "minimum": 0.0,
            "title": "Market Min"
          },
          "market_max": {
            "type": "number",
            "minimum": 0.0,
            "title": "Market Max"
          },
          "recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recycler Id"
          },
          "source": {
            "type": "string",
            "pattern": "^(recycler_quote|field_survey|synthetic)$",
            "title": "Source",
            "default": "synthetic"
          }
        },
        "type": "object",
        "required": [
          "category",
          "district",
          "latitude",
          "longitude",
          "buying_price",
          "selling_quoted_price",
          "market_min",
          "market_max"
        ],
        "title": "PriceCreate"
      },
      "PriceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "sub_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Category"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "buying_price": {
            "type": "number",
            "title": "Buying Price"
          },
          "selling_quoted_price": {
            "type": "number",
            "title": "Selling Quoted Price"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "market_min": {
            "type": "number",
            "title": "Market Min"
          },
          "market_max": {
            "type": "number",
            "title": "Market Max"
          },
          "recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recycler Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "sub_category",
          "district",
          "city",
          "latitude",
          "longitude",
          "recorded_at",
          "buying_price",
          "selling_quoted_price",
          "unit",
          "market_min",
          "market_max",
          "recycler_id",
          "source",
          "created_at"
        ],
        "title": "PriceResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}